Update upstream documentaiton after 5.8, rebuild
This change rebuilds the website, containing the latest upstream
documentation (including everything in 5.8, and probably everything in
the 5.9 merge window).

It also rebuilds almost everything with a slightly newer version of
sphinx (2.4.3). Note that the "stable kernel" documentation wasn't
rebuilt, as the newer sphinx version cannot build such an old kernel
(some deprecated stuff was removed).

Going forward, it may be worth just linking to kernel.org for the
doucmentation, and getting rid of the stable kernel stuff entirely, but
this will do for now.

Signed-off-by: David Gow <davidgow@google.com>